consolidate and add test
