+++ src/amg/azext_amg/restore.py
@@ -1,16 +1,14 @@
+import collections
+import json
 from glob import glob
 import tarfile
 import tempfile
-import collections
 
 from azure.cli.core.azclierror import ArgumentUsageError
 from knack.log import get_logger
 
-from .create_folder import create_folder
-from .create_dashboard import create_dashboard
-from .create_snapshot import create_snapshot
-from .create_annotation import create_annotation
-from .create_datasource import create_datasource
+from .utils import get_folder_id, send_grafana_post
+
 
 logger = get_logger(__name__)
 
@@ -24,20 +22,20 @@ def restore(grafana_url, archive_file, components, http_headers):
     # Shell game magic warning: restore_function keys require the 's'
     # to be removed in order to match file extension names...
     restore_functions = collections.OrderedDict()
-    restore_functions['folder'] = create_folder
-    restore_functions['dashboard'] = create_dashboard
-    restore_functions['snapshot'] = create_snapshot
-    restore_functions['annotation'] = create_annotation
-    restore_functions['datasource'] = create_datasource
+    restore_functions['folder'] = _create_folder
+    restore_functions['dashboard'] = _create_dashboard
+    restore_functions['snapshot'] = _create_snapshot
+    restore_functions['annotation'] = _create_annotation
+    restore_functions['datasource'] = _create_datasource
 
     with tarfile.open(name=archive_file, mode='r:gz') as tar:
         with tempfile.TemporaryDirectory() as tmpdir:
             tar.extractall(tmpdir)
             tar.close()
-            restore_components(grafana_url, restore_functions, tmpdir, components, http_headers)
+            _restore_components(grafana_url, restore_functions, tmpdir, components, http_headers)
 
 
-def restore_components(grafana_url, restore_functions, tmpdir, components, http_headers):
+def _restore_components(grafana_url, restore_functions, tmpdir, components, http_headers):
 
     if components:
         exts = [c[:-1] for c in components]
@@ -50,3 +48,72 @@ def restore_components(grafana_url, restore_functions, tmpdir, components, http_
         for file_path in glob('{0}/**/*.{1}'.format(tmpdir, ext), recursive=True):
             logger.info('Restoring %s: %s', ext, file_path)
             restore_functions[ext](grafana_url, file_path, http_headers)
+
+
+# Restore dashboards
+def _create_dashboard(grafana_url, file_path, http_headers):
+    with open(file_path, 'r', encoding="utf8") as f:
+        data = f.read()
+
+    content = json.loads(data)
+    content['dashboard']['id'] = None
+
+    payload = {
+        'dashboard': content['dashboard'],
+        'folderId': get_folder_id(content, grafana_url, http_post_headers=http_headers),
+        'overwrite': True
+    }
+
+    result = send_grafana_post('{0}/api/dashboards/db'.format(grafana_url), json.dumps(payload), http_headers)
+    dashboard_title = content['dashboard'].get('title', '')
+    logger.warning("Create dashboard \"%s\". %s", dashboard_title, "SUCCESS" if result[0] == 200 else "FAILURE")
+    logger.info("status: %s, msg: %s", result[0], result[1])
+
+
+# Restore snapshots
+def _create_snapshot(grafana_url, file_path, http_headers):
+    with open(file_path, 'r', encoding="utf8") as f:
+        data = f.read()
+
+    snapshot = json.loads(data)
+    try:
+        snapshot['name'] = snapshot['dashboard']['title']
+    except KeyError:
+        snapshot['name'] = "Untitled Snapshot"
+
+    (status, content) = send_grafana_post('{0}/api/snapshots'.format(grafana_url), json.dumps(snapshot), http_headers)
+    logger.warning("Create snapshot \"%s\". %s", snapshot['name'], "SUCCESS" if status == 200 else "FAILURE")
+    logger.info("status: %s, msg: %s", status, content)
+
+
+# Restore folders
+def _create_folder(grafana_url, file_path, http_headers):
+    with open(file_path, 'r', encoding="utf8") as f:
+        data = f.read()
+
+    folder = json.loads(data)
+    result = send_grafana_post('{0}/api/folders'.format(grafana_url), json.dumps(folder), http_headers)
+    logger.warning("Create folder \"%s\". %s", folder.get('title', ''), "SUCCESS" if result[0] == 200 else "FAILURE")
+    logger.info("status: %s, msg: %s", result[0], result[1])
+
+
+# Restore annotations
+def _create_annotation(grafana_url, file_path, http_headers):
+    with open(file_path, 'r', encoding="utf8") as f:
+        data = f.read()
+
+    annotation = json.loads(data)
+    result = send_grafana_post('{0}/api/annotations'.format(grafana_url), json.dumps(annotation), http_headers)
+    logger.warning("Create annotation \"%s\". %s", annotation['id'], "SUCCESS" if result[0] == 200 else "FAILURE")
+    logger.info("status: %s, msg: %s", result[0], result[1])
+
+
+# Restore data sources
+def _create_datasource(grafana_url, file_path, http_headers):
+    with open(file_path, 'r', encoding="utf8") as f:
+        data = f.read()
+
+    datasource = json.loads(data)
+    result = send_grafana_post('{0}/api/datasources'.format(grafana_url), json.dumps(datasource), http_headers)
+    logger.warning("Create datasource \"%s\". %s", datasource['name'], "SUCCESS" if result[0] == 200 else "FAILURE")
+    logger.info("status: %s, msg: %s", result[0], result[1])
