+++ src/amg/azext_amg/tests/latest/test_amg_scenario.py
@@ -4,7 +4,9 @@
 # --------------------------------------------------------------------------------------------
 
 import os
+import tempfile
 import unittest
+
 
 from azure.cli.testsdk import (ScenarioTest, ResourceGroupPreparer, MSGraphNameReplacer, MOCKED_USER_NAME)
 from azure.cli.testsdk .scenario_tests import AllowLargeResponse
@@ -283,6 +285,84 @@ class AmgScenarioTest(ScenarioTest):
             self.cmd('grafana delete -g {rg} -n {name} --yes')
             final_count = len(self.cmd('grafana list').get_output_in_json())
             self.assertTrue(final_count, count - 1)
+
+
+    @AllowLargeResponse(size_kb=3072)
+    @ResourceGroupPreparer(name_prefix='cli_test_amg', location='westcentralus')
+    def test_amg_backup_restore(self, resource_group):
+
+        # Test Instance
+        self.kwargs.update({
+            'name': 'clitestamg',
+            'location': 'westcentralus',
+            'name2': 'clitestamg2',
+            'tempDir': tempfile.TemporaryDirectory()
+        })
+
+        owner = self._get_signed_in_user()
+        self.recording_processors.append(MSGraphNameReplacer(owner, MOCKED_USER_NAME))
+
+        with unittest.mock.patch('azext_amg.custom._gen_guid', side_effect=self.create_guid):
+
+            amg1 = self.cmd('grafana create -g {rg} -n {name} -l {location}').get_output_in_json()
+
+            # set up folder
+            self.kwargs.update({
+                'folderTitle': 'Test Folder',
+                'id': amg1['id']
+            })
+            self.cmd('grafana folder create -g {rg} -n {name} --title "{folderTitle}"')
+            
+            # set up data source
+            self.kwargs.update({
+                'dataSourceDefinition': test_data_source,
+                'dataSourceName': test_data_source["name"]
+            })
+            self.cmd('grafana data-source create -g {rg} -n {name} --definition "{dataSourceDefinition}"')        
+        
+            # create dashboard
+            dashboard_title = test_dashboard["dashboard"]["title"]
+            slug = dashboard_title.lower().replace(' ', '-')
+
+            self.kwargs.update({
+                'dashboardDefinition': test_dashboard,
+                'dashboardTitle': dashboard_title,
+                'dashboardSlug': slug,
+            })
+            response_create = self.cmd('grafana dashboard create -g {rg} -n {name} --folder {folderTitle}  --definition "{dashboardDefinition}" --title "{dashboardTitle}"').get_output_in_json()
+
+            self.kwargs.update({
+                'dashboardUid': response_create["uid"],
+            })
+
+            self.cmd('grafana backup -g {rg} -n {name} -d {tempDir}')
+
+            filenames = next(os.walk(tempDir), (None, None, []))[2]
+            self.assertTrue(len(filenames) == 1)
+            self.assertTrue(filenames[0].endswith('.tar.gz'))
+
+            self.kwargs.update({
+                'archiveFile': os.path.join(tempDir, filenames[0])
+            })
+
+            self.cmd('grafana folder delete -g {rg} -n {name} --folder "{folderTitle}"')
+            self.cmd('grafana data-source delete -g {rg} -n {name} --data-source "{dataSourceName}"') 
+
+            self.cmd('grafana restore -g {rg} -n {name} --archive-file {archiveFile}')
+
+            self.cmd('grafana data-source show -g {rg} -n {name} --data-source "{dataSourceName}')
+            self.cmd('grafana folder show -g {rg} -n {name} --folder "{folderTitle}"')
+            self.cmd('grafana dashboard show -g {rg} -n {name} --dashboard "{dashboardUid}"')  # add a bit check
+
+            amg2 = self.cmd('grafana create -g {rg} -n {name2} -l {location}').get_output_in_json()
+            self.kwargs.update({
+                'id2': amg2['id']
+            })
+
+            self.cmd('grafana dashboard sync --source {id} --destination {id2} --folders-to-include {folderTitle}')
+            self.cmd('grafana data-source show -g {rg} -n {name2} --data-source "{dataSourceName}')
+            self.cmd('grafana folder show -g {rg} -n {name2} --folder "{folderTitle}"')
+            self.cmd('grafana dashboard show -g {rg} -n {name2} --dashboard "{dashboardUid}"')  # add a bit check
 
 
     def _get_signed_in_user(self):
